+++ Pages/Medications/Details.cshtml.cs
@@ -1,22 +1,18 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using Roshta.Data;
 using Roshta.Models;
+using Roshta.Repositories.Interfaces;
+using System.Threading.Tasks;
 
 namespace Roshta.Pages_Medications
 {
     public class DetailsModel : PageModel
     {
-        private readonly Roshta.Data.ApplicationDbContext _context;
+        private readonly IMedicationRepository _medicationRepository;
 
-        public DetailsModel(Roshta.Data.ApplicationDbContext context)
+        public DetailsModel(IMedicationRepository medicationRepository)
         {
-            _context = context;
+            _medicationRepository = medicationRepository;
         }
 
         public Medication Medication { get; set; } = default!;
@@ -28,16 +24,16 @@ namespace Roshta.Pages_Medications
                 return NotFound();
             }
 
-            var medication = await _context.Medications.FirstOrDefaultAsync(m => m.Id == id);
-
-            if (medication is not null)
+            var medication = await _medicationRepository.GetByIdAsync(id.Value);
+            if (medication == null)
+            {
+                return NotFound();
+            }
+            else
             {
                 Medication = medication;
-
-                return Page();
             }
-
-            return NotFound();
+            return Page();
         }
     }
 }